[REFACTOR]: ExpiredCouponDto 위치 변경
