+++ project/stampee/src/main/java/dto/response/ExpiredCouponDto.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.response;
 
 public class ExpiredCouponDto {
 	private String username;
