+++ project/stampee/src/test/java/service/CouponServiceTest.java
@@ -15,7 +15,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import dto.ExpiredCouponDto;
+import dto.response.ExpiredCouponDto;
 import repository.CouponRepository;
 
 @ExtendWith(MockitoExtension.class)
